This check-in includes the following feature enhancements:
Conversion to SAM:

    Added a wrapper script called east.sh.  This is now the preferred
    top-level executable file. This shell script now contains the
    various steps that is required to convert ACE file to SAM file
    using AMOS tools. The GUI provides a -CONVERT_TO_SAM command line
    for this script to do the conversion. The GUI generates the
    necessary command-line arguments for conversion to kick-in.

	Unfortunately, on Red Hawk AMOS tools don't work on compute nodes
	and consequently ACE-to-SAM conversion fails. Hopefully Jens will
	look into it.


Progress Generation:

    Now EAST generates some coarse progress information just like
    PEACE. So on long running jobs the user will get some feedback on
    the job progress.  However, the progress information is very
	coarse just reporting major steps (such as: 6-tuple forming,
	6-tuple processing, and contig generation)

+++ EAST/C++/OvlDistance.cpp
@@ -192,7 +192,7 @@ vector<int> OvlDistance::getOVLDistance(const string& tS1, const string& tS2, bo
 	 *
 	 * Note: if ovlWindowSize > length of s1, disRight and disLeft would not be equal to zero.
 	 */
-	if ((abs(lenOverlap) == s1.length()) && (s1.length() <= s2.length())
+	if ((abs(lenOverlap) == (int) s1.length()) && (s1.length() <= s2.length())
 			&& ((disRight <= InclusionThreshold) || (abs(disLeft)
 					<= InclusionThreshold))) {
 		lenOverlap = min(s1.length(), s2.length());
@@ -347,7 +347,7 @@ bool OvlDistance::checkInclusion(const string& s1, const string& s2, bool create
 
 	/*if s1 is included in s2, return true, else return false
 	 */
-	if ((abs(lenOverlap) == s1.length()) && (s1.length() <= s2.length())
+	if ((abs(lenOverlap) == (int) s1.length()) && (s1.length() <= s2.length())
 			&& ((disRight <= InclusionThreshold) || (abs(disLeft)
 					<= InclusionThreshold))) {
 		return true;
